requirements.txt
PyQt6
pycryptodome
pathlib
rsa
overrides